\name{ANN2-package}
\alias{ANN2-package}
\alias{ANN2}
\docType{package}
\title{
\packageTitle{ANN2}
}
\description{
\packageDescription{ANN2}
}
\details{
Package for training neural networks. Special options for detecting and plotting 
anomalies using autoencoding neural networks. 
}
\author{
\packageAuthor{ANN2}

Maintainer: \packageMaintainer{ANN2}
}
\references{
Add links to references. Efficient Backprop Le Cun
}
\keyword{ package }