% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compression_plot}
\alias{compression_plot}
\alias{compression_plot.ANN}
\title{Compression plot}
\usage{
compression_plot(object, ...)

\method{compression_plot}{ANN}(object, X, colors = NULL, ...)
}
\arguments{
\item{object}{autoencoder object of class \code{ANN}}

\item{\dots}{arguments to be passed down}

\item{X}{data matrix with original values to be compressed and plotted}

\item{colors}{optional vector of discrete colors}
}
\value{
Plots
}
\description{
plot compressed observation in pairwise dimensions
}
\details{
Matrix plot of pairwise dimensions
}
