% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{comprPlot}
\alias{comprPlot}
\title{Compression plot}
\usage{
comprPlot(object, X, colors = NULL)
}
\arguments{
\item{object}{autoencoder object of class \code{ANN}}

\item{X}{data matrix with original values to be compressed and plotted}

\item{colors}{optional vector of discrete colors}
}
\value{
Plots
}
\description{
\code{comprPlot} plot compressed observation in pairwise dimensions
}
\details{
Matrix plot of pairwise dimensions
}
