% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{decode}
\alias{decode}
\title{Decoding step}
\usage{
decode(object, compressed)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{compressed}{Data to decompress}
}
\description{
Decompress low-dimensional representation resulting from the nodes
of the middle layer. Output are the reconstructed inputs to function \code{encode}
}
