% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{encode}
\alias{encode}
\title{Encoding step}
\usage{
encode(object, newdata, returnInputs = FALSE)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{newdata}{Data to compress}

\item{returnInputs}{Logical indicating whether the inputs to the compression
layer should be returned. Mainly used for plotting the activations.}
}
\description{
Compress data according to trained replicator or autoencoder.
Outputs are the activations of the nodes in the middle layer for each 
observation in \code{newdata}
}
