% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ANN}
\alias{plot.ANN}
\title{Plot training and validation loss}
\usage{
\method{plot}{ANN}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ANN}}

\item{...}{further arguments to be passed to plot}
}
\value{
Plots
}
\description{
\code{plot} Generate plots of the loss against epochs
}
\details{
A genereric function for training neural nets
}
