% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{plot.rANN}
\alias{plot.rANN}
\title{Plot mahalanobis distances of reconstructed errors}
\usage{
\method{plot}{rANN}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{Object of class \code{rX}}

\item{alpha}{significance level for determining cut-off point of squared
Mahalanobis distanced using the chi-square distribution}

\item{...}{further arguments to be passed to plot}
}
\value{
Plots
}
\description{
\code{plot} Generate plots of the mahalanobis distances for each reconstructed observation
}
\details{
A genereric function for training neural nets
}
