% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{plotStepFunction}
\alias{plotStepFunction}
\title{Plot the step function used in \code{replicator}}
\usage{
plotStepFunction(object = NULL, X, hidden_node = 1, color = "red",
  derivative = FALSE, ...)
}
\arguments{
\item{object}{Trained replicator neural network object}

\item{X}{Data to be plotted as points on steps}

\item{hidden_node}{Number of the node to plot}

\item{color}{Color of points. Replace with a vector indicating class-membership
to give points in classes different colors.}

\item{derivative}{logical indicating if the derivative of the step function should be plotted.}

\item{...}{further arguments to be passed to plot}
}
\description{
Plot the projections of observations onto the stepfunction in the low-dimensional 
space of the compression layer
}
