% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{predict.ANN}
\alias{predict.ANN}
\title{Make predictions for new data}
\usage{
\method{predict}{ANN}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{newdata}{Data to make predictions on}

\item{...}{further arguments (not in use)}
}
\value{
A list with predicted classes for classification and fitted probabilities
}
\description{
\code{predict} Predict class or value for new data
}
\details{
A genereric function for training neural nets
}
