% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.ANN}
\alias{print.ANN}
\title{Print ANN}
\usage{
\method{print}{ANN}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ANN}}

\item{\dots}{Further arguments}
}
\description{
Print info on trained Neural Network
}
