% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{read_ANN}
\alias{read_ANN}
\title{Read ANN object from file}
\usage{
read_ANN(file)
}
\arguments{
\item{file}{character specifying file path}
}
\value{
Object of class ANN
}
\description{
Deserialize ANN object from binary file
}
