% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct data using trained ANN object of type autoencoder}
\usage{
reconstruct(object, X)
}
\arguments{
\item{object}{Object of class \code{ANN} created with \code{autoencoder()}}

\item{X}{data matrix to reconstruct}
}
\value{
Reconstructed observations and anomaly scores (reconstruction errors)
}
\description{
\code{reconstruct} takes new data as input and reconstructs the observations using
a trained replicator or autoencoder object.
}
\details{
A genereric function for training neural nets
}
