% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{reconstruction_plot}
\alias{reconstruction_plot}
\alias{reconstruction_plot.ANN}
\title{Reconstruction plot}
\usage{
reconstruction_plot(object, ...)

\method{reconstruction_plot}{ANN}(object, X, colors = NULL, ...)
}
\arguments{
\item{object}{autoencoder object of class \code{ANN}}

\item{\dots}{arguments to be passed down}

\item{X}{data matrix with original values to be reconstructed and plotted}

\item{colors}{optional vector of discrete colors. The reconstruction errors
are are used as color if this argument is not specified}
}
\value{
Plots
}
\description{
plots original and reconstructed data points in a single plot with connecting 
lines between original value and corresponding reconstruction
}
\details{
Matrix plot of pairwise dimensions
}
