% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{recPlot}
\alias{recPlot}
\title{Reconstruction plot}
\usage{
recPlot(object, X, colors = NULL)
}
\arguments{
\item{object}{autoencoder object of class \code{ANN}}

\item{X}{data matrix with original values to be reconstructed and plotted}

\item{colors}{optional vector of discrete colors. The reconstruction errors
are are used as color if this argument is not specified}
}
\value{
Plots
}
\description{
\code{recPlot} plot original and reconstructed data points in a single plot 
with connecting lines between original value and corresponding reconstruction
}
\details{
Matrix plot of pairwise dimensions
}
