% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setActivParams}
\alias{setActivParams}
\title{Check user input related to activation functions}
\usage{
setActivParams(activ.functions, step.H, step.k, meta)
}
\description{
Set and check activation parameter list
}
\keyword{internal}
