% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setLossParams}
\alias{setLossParams}
\title{Check user input related to loss function}
\usage{
setLossParams(loss.type, huber.delta, meta)
}
\description{
Set and check loss parameter list
}
\keyword{internal}
