% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setNetworkParams}
\alias{setNetworkParams}
\title{Check user input related to network structure}
\usage{
setNetworkParams(hidden.layers, standardize, verbose, meta)
}
\description{
Set and check network parameter list
}
\keyword{internal}
