% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setOptimParams}
\alias{setOptimParams}
\title{Check user input related to optimizer}
\usage{
setOptimParams(optim.type, learn.rates, L1, L2, sgd.momentum,
  rmsprop.decay, adam.beta1, adam.beta2, meta)
}
\description{
Set and check optimizer parameter list
}
\keyword{internal}
