% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{setTrainParams}
\alias{setTrainParams}
\title{Check user input related to training}
\usage{
setTrainParams(n.epochs, batch.size, val.prop, drop.last, random.seed,
  data)
}
\description{
Set training parameters
}
\keyword{internal}
