% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{write_ANN}
\alias{write_ANN}
\title{Write ANN object to file}
\usage{
write_ANN(object, file)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{file}{character specifying file path}
}
\description{
Serialize ANN object to binary file
}
