\name{ANN2-package}
\alias{ANN2-package}
\alias{ANN2}
\docType{package}
\title{
\packageTitle{ANN2}
}
\description{
\packageDescription{ANN2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ANN2}
\packageIndices{ANN2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ANN2}

Maintainer: \packageMaintainer{ANN2}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
