% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{decode}
\alias{decode}
\alias{decode.ANN}
\title{Decoding step}
\usage{
decode(object, ...)

\method{decode}{ANN}(object, compressed, compression.layer = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{\dots}{arguments to be passed down}

\item{compressed}{Compressed data}

\item{compression.layer}{Integer specifying which hidden layer is the 
compression layer. If NULL this parameter is inferred from the structure 
of the network (hidden layer with smallest number of nodes)}
}
\description{
Decompress low-dimensional representation resulting from the nodes
of the middle layer. Output are the reconstructed inputs to function \code{encode()}
}
