% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{encode}
\alias{encode}
\alias{encode.ANN}
\title{Encoding step}
\usage{
encode(object, ...)

\method{encode}{ANN}(object, newdata, compression.layer = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{ANN}}

\item{\dots}{arguments to be passed down}

\item{newdata}{Data to compress}

\item{compression.layer}{Integer specifying which hidden layer is the 
compression layer. If NULL this parameter is inferred from the structure 
of the network (hidden layer with smallest number of nodes)}
}
\description{
Compress data according to trained replicator or autoencoder.
Outputs are the activations of the nodes in the middle layer for each 
observation in \code{newdata}
}
