% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ANN}
\alias{plot.ANN}
\title{Plot training and validation loss}
\usage{
\method{plot}{ANN}(x, max.points = 1000, ...)
}
\arguments{
\item{x}{Object of class \code{ANN}}

\item{max.points}{Maximum number of points to plot, set to NA, NULL or Inf to
include all points in the plot}

\item{...}{further arguments to be passed to plot}
}
\value{
Plots
}
\description{
\code{plot} Generate plots of the loss against epochs
}
\details{
A generic function for plot loss of neural net
}
