% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{plotStepFunction}
\alias{plotStepFunction}
\title{Plot the step function used in \code{replicator}}
\usage{
plotStepFunction(nSteps = 10, smoothSteps = 100, derivative = FALSE, ...)
}
\arguments{
\item{nSteps}{number of steps on interval (0, 1)}

\item{smoothSteps}{smoothing parameter. Small values results in high smoothness.}

\item{derivative}{logical indicating if the derivative of the step function should be plotted.}

\item{...}{further arguments to be passed to plot}
}
\description{
Change parameters to see how the shape of the step function is affected.
Note that setting smoothSteps very high results in a large deriative,
this might cause instability during training.
}
