% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callable.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct data using trained Autoencoder or Replicator object}
\usage{
reconstruct(NN, X, mahalanobis = TRUE)
}
\arguments{
\item{NN}{Object of class \code{NN}}

\item{X}{data (matrix) to reconstruct}

\item{mahalanobis}{logical indicating if Mahalanobis distance should be calculated}
}
\value{
Reconstructed observations and optional Mahalanobis distances
}
\description{
\code{reconstruct} takes new data as input and reconstructs the observations using
a trained replicator or autoencoder object.
}
\details{
A genereric function for training neural nets
}
